+++ app_configs/__init__.py
@@ -2,6 +2,7 @@ from dir_maker import Config as Directories
 from user_settings import Config as UserConfig
 from keybinds import Config as KeybindConfig
 from sidebar_loader import Config as SidebarLoader
+from terminal_input_handler import Config as TerminalConfig
 
 
 class AppConfigHandler(object):
@@ -11,7 +12,8 @@ class AppConfigHandler(object):
         'directories': Directories(),
         'user_settings': UserConfig(),
         'keybinder': KeybindConfig(),
-        'sidebar_loader': SidebarLoader()
+        'sidebar_loader': SidebarLoader(),
+        'terminal_config': TerminalConfig()
     }
 
     def __init__(self, root_widget):