+++ app_modules/widgets_integrated/labelbutton.py
@@ -2,13 +2,14 @@ from kivy.lang import Builder
 from kivy.uix.label import Label
 from kivy.uix.behaviors import ButtonBehavior
 from kivy.properties import ListProperty
-from app_modules.widgets.multi_line_label import MultiLineLabel
+from app_modules.widgets_standalone.multi_line_label import MultiLineLabel
+
 
 class LabelButton(ButtonBehavior, MultiLineLabel):
     pass
 
 
-bgkv = """
+kv = """
 <LabelButtonBackground>
     canvas.before:
         Color:
@@ -20,9 +21,10 @@ bgkv = """
 
 class LabelButtonBackground(ButtonBehavior, Label):
     background_color = ListProperty([1,1,1,1])
+
     def __init__(self, **kwargs):
         super(LabelButtonBackground, self).__init__(**kwargs)
 
 
 
-Builder.load_string(bgkv)
+Builder.load_string(kv)